version presentacion

+++ Front/src/app/list-atributes/list-atributes.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ApiService } from '../api.service';
 import { Atributo } from '../Interfaces/Atributo';
+import Swal from 'sweetalert2'
+
 
 @Component({
   selector: 'app-list-atributes',
@@ -19,9 +21,31 @@ export class ListAtributesComponent implements OnInit {
     this._api.listAtributes().subscribe(objeto=>this.atributes=objeto);
   }
   deleteAtribute(e:string){
-    this._api.deleteAtribute(e);
-    this.listAtributes();
-    this.router.navigate(['list-atributes']);
+
+    Swal.fire({
+      title: '¿Seguro que quiere borrar el atributo?',
+      text: "no podrá deshacer este paso",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar'
+    }).then((result) => {
+  
+      if (result.isConfirmed) {
+        console.log(e);
+        this._api.deleteAtribute(e);
+        this.listAtributes();
+        this.router.navigate(['']);
+        Swal.fire(
+          'Borrado',
+          'El atributo ha sido borrado',
+          'success'
+        )
+      }
+    })
+    
+ 
     
 
   }
